feat: Added educational file template.
